services/llm-server/requirements.txt
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
